requirements.txt
streamlit
torch
transformers
sentence-transformers
faiss-cpu
pillow
numpy
ollama
